+++ app/src/main/res/layout/activity_program_mode.xml
@@ -24,7 +24,6 @@
             android:layout_marginLeft="@dimen/rocket_image_margin_start"
             android:layout_marginEnd="@dimen/rocket_image_margin_end"
             android:layout_marginRight="@dimen/rocket_image_margin_end"
-            android:onClick="moveRocket"
             android:src="@drawable/ic_rocket_image" />
 
     </RelativeLayout>
